Initial Python script to generate KDE apidox
Currently produces something resembling the correct output if you run it
on a framework like KCoreAddons.  Implements almost all of that part of
the functionality of doxygen.sh, only it actually works with Qt 5 and
the framework directory structure.  See the FIXME notes for missing
functionality.

It would be good to have another mode/script that would generate a
bundle of documentation, generating all the ones for the frameworks
(including ordering by tier, etc).  This is what doxygen.sh used to do
on kdelibs.