Use "svn export" to fetch KDE identities if possible, and cache them
WebSVN is not set up to deal with a large number of hits - it is
intended for casual browsing. "svn export" is much more efficient, and
the anonsvn servers are meant to take this load.

Given that we would like developers to be generating documentation to
check their doxygen markup, we should use a method that scales.

This uses "svn export" (either through libsvn python bindings or the
command-line utility) if it can. It also caches the result for a day,
since account information does not change that often (and being
up-to-date is not that important for developers generating the dox).

REVIEW: 119805