Show q_properties at the top of class documentation
This brings it in line with Qt, and makes it hopefully a bit easier for
QML users to find things.

Original doxygenlayout xml is the template doxygen file but with
properties moved.

Longer term I want to make it autopopulate @accessor methods so it jumps to the relevant bit of doc.

REVIEW: 119723