Kapidox generates qch files
Summary:
Hey guys,

could you check if the qch generation is usable for you please?

for this, please run

`path/to/kapidox/src/kapidox_generate path/to/your/source/files --qhp`

And chech in the `qch/` folder that it suits you.

(The code will need refactoring, but first I want to be sure it does something good)

Reviewers: kossebau, #kdevelop!, mwolff, apol

Subscribers: antonanikin, #frameworks, #documentation

Tags: #frameworks, #documentation

Differential Revision: https://phabricator.kde.org/D5275