Look first for qhelpgenerator-qt5 for help generation
Summary:
A number of distributions uses the -qt5 suffix for Qt command line
tools, to make sure they can coexist with the Qt4 equivalents.

This change uses `distutils`'s `find_executable` to look for the -qt5
suffix of `qhelpgenerator`. If it is not found, it falls back to the
un-suffixed version.

Reviewers: #frameworks, ochurlaud

Reviewed By: ochurlaud

Subscribers: fvogt, #documentation

Tags: #frameworks, #documentation

Differential Revision: https://phabricator.kde.org/D11315