Update IRC network name in HTML templates
KDE has moved from freenode to Libera Chat. The HTML templates link to the
project's IRC channel on the documentation sidebar if there is a channel
set in the metadata, but the server is hardcoded to "freenode". This commit
replaces it with Libera Chat, and uses the full hostname instead of just
the name (I'm not sure how that ever worked before).