Preprocessing: avoid .. as name
One (common) footcannon with kapidox is running, say,
	kapidox_generate ..
like one would run cmake -- thinking that that will build the dox
of the directory above. This footcannon does three things:
 - sets the name of the project being built to ".."
 - builds the documentation in ".."
 - rm -rf ..
The last item is the most footcannon. Avoid this by calculating
the real name (use abspath, even though it's expensive) of the
path, so that even building docs inside a source directory will
put the docs **below** the docs-build directory and not remove
all of the sources as a side-effect.