modify kapidox for docker deployment
This proof of concept shows how the tools could be deployed in a docker container.

It is based on the idea that we want to make it possible to run on the CI infrastructure as a docker container, but also provide a way to develop and test outside of docker (see bootstrap-devenv.sh).

To make this possible we do not need to build a package. An editable install provides the same functionality with a less complex setup.

We considered the introduction of third party tools like e.g. poetry as proposed in !11 or tox, or pipenv, or flit, etc., but it seemed more feasible to simplify the existing structure and remove everything that is not absolutely necessary for functioning. This includes all setup around creating a distribution package and also the removal of the extra src folder which is not necessary in the current setup.

This is not quite there yet, but before investing more work, we would like to invite comments if this is a feasible direction.

One of the things still missing, is that the input and output folders need to be mapped into the container. There are several ways to do that (binds mounts, volumes, ...), this can be discussed and tested if we decide to move on in this direction.

Co-Authored-By: Frederik Schwarzer <schwarzer@kde.org>